Edited Parser and Created some test cases on the parser
I edited each of ParsingTable class and used unordered_map for the table instead of the 2d vectors.
I tried to edit the ParsingTable in the generator class but i found that we face an infinite loop during running the test case (test_generation) i provided in the testing.cpp file.
I did not remove any code but just comment it.